#191: address PR review — repackage stragglers, drop AppContainer.inboundEvents, rename banner
- move NoOpConnectionMonitor → transfer/, next to ConnectionMonitor interface
- move DeviceNameRepublisher → discovery/, next to MdnsDiscovery
- drop AppContainer.inboundEvents — RootComponentFactory already defaults inboundEvents to MutableSharedFlow()
- rename IosForegroundConstraintBanner → ForegroundConstraintBanner (commonMain file can't carry platform name)
